+++ spec/rubocop/cop/style/ascii_comments_spec.rb
@@ -11,6 +11,17 @@ describe RuboCop::Cop::Style::AsciiComments do
                    ['# encoding: utf-8',
                     '# 这是什么？'])
     expect(cop.offenses.size).to eq(1)
+    expect(cop.highlights).to eq(['这是什么？'])
+    expect(cop.messages)
+      .to eq(['Use only ascii symbols in comments.'])
+  end
+
+  it 'registers an offense for commentes with mixed chars' do
+    inspect_source(cop,
+                   ['# encoding: utf-8',
+                    '# foo ∂ bar'])
+    expect(cop.offenses.size).to eq(1)
+    expect(cop.highlights).to eq(['∂'])
     expect(cop.messages)
       .to eq(['Use only ascii symbols in comments.'])
   end